Expose `trait KeyExchange` publicly
