+++ src/messages.rs
@@ -25,7 +25,7 @@ use crate::errors::ProtocolError;
 use crate::hash::OutputSize;
 use crate::key_exchange::group::Group;
 use crate::key_exchange::shared::NonceLen;
-use crate::key_exchange::traits::{
+use crate::key_exchange::{
     Deserialize, Ke1MessageLen, Ke2MessageLen, Ke3MessageLen, KeyExchange, Serialize,
     SerializedCredentialRequest, SerializedCredentialResponse,
 };
